reorganized vue components in folders
