+++ Components/MapWidgets/LegendGUI.vue
