+++ Components/MapWidgets/WMSLegend.vue
