+++ Components/MapWidgets/WidgetCombinedRadars.vue
@@ -50,7 +50,7 @@
 
 // Import components
 import LegendGUI from "./LegendGUI.vue";
-import OnOffButton from "./OnOffButton.vue";
+import OnOffButton from "../OnOffButton.vue";
 
 export default {
   name: 'widgetCombinedRadars', // Caps, no -
