+++ Components/MapWidgets/WidgetWeatherLayers.vue
